fix: 修复数据库类型转换与PostgreSQL兼容性问题
将多处SQL查询中的CAST(xxx AS INTEGER)调整为BIGINT以避免大额数值溢出，同时修正generate_transaction_no函数的时间查询语句适配PostgreSQL

+++ src-tauri/src/commands/inventory.rs
@@ -402,7 +402,7 @@ pub async fn get_inventory_list(
                inv.warehouse_id, w.name AS warehouse_name,
                inv.quantity, inv.reserved_qty, inv.available_qty,
                inv.avg_cost,
-               CAST(ROUND((inv.quantity * inv.avg_cost)::numeric, 0) AS INTEGER) AS inventory_value,
+               CAST(ROUND((inv.quantity * inv.avg_cost)::numeric, 0) AS BIGINT) AS inventory_value,
                m.safety_stock, m.max_stock,
                CASE
                    WHEN m.safety_stock IS NOT NULL AND inv.available_qty < m.safety_stock THEN 'low'
@@ -551,7 +551,7 @@ pub async fn get_inventory_detail(
         SELECT inv.warehouse_id, w.name AS warehouse_name,
                inv.quantity, inv.reserved_qty, inv.available_qty,
                inv.avg_cost,
-               CAST(ROUND((inv.quantity * inv.avg_cost)::numeric, 0) AS INTEGER) AS inventory_value,
+               CAST(ROUND((inv.quantity * inv.avg_cost)::numeric, 0) AS BIGINT) AS inventory_value,
                inv.last_in_date, inv.last_out_date
         FROM inventory inv
         JOIN warehouses w ON w.id = inv.warehouse_id
